feat: update project configuration and add CORS middleware with logging support
