+++ functions/.eslintrc.cjs
@@ -18,8 +18,7 @@ module.exports = {
     sourceType: "module",
   },
   ignorePatterns: [
-    "/lib/**/*", // Ignore built files.
-    "/generated/**/*", // Ignore generated files.
+    "/lib/**/*",
   ],
   plugins: [
     "@typescript-eslint",
@@ -29,5 +28,7 @@ module.exports = {
     "quotes": ["error", "double"],
     "import/no-unresolved": 0,
     "indent": ["error", 2],
+    "object-curly-spacing": ["error", "never"],
+    "max-len": ["error", {"code": 120}],
   },
 };
