+++ functions/package.json
@@ -14,7 +14,9 @@
     "node": "18"
   },
   "main": "lib/index.js",
+  "type": "module",
   "dependencies": {
+    "chalk": "^5.4.1",
     "cors": "^2.8.5",
     "firebase-admin": "^11.8.0",
     "firebase-functions": "^6.3.2",
