+++ functions/src/config.ts
@@ -1,3 +1,3 @@
 export const RESEND_API_KEY = "re_XsCu6Gft_94MGXrHPLzL6nuwSvbzhiQ4S";
-export const EMAIL_FROM = "contact@bodysharing.com";
+export const EMAIL_FROM = "onboarding@resend.dev";
 export const EMAIL_ADMIN = "yan.braslavsky@gmail.com";
